Delete course fix

+++ src/components/authors/AuthorsPage.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import * as authorActions from '../../redux/actions/authorActions';
+import * as coursesActions from '../../redux/actions/courseActions';
 import PropTypes from 'prop-types';
 import Spinner from '../common/Spinner';
 import { Redirect } from 'react-router-dom';
@@ -9,6 +10,7 @@ import { toast } from 'react-toastify';
 import AuthorList from './AuthorList';
 
 const AuthorsPage = ({
+  courses,
   authors,
   loading,
   actions 
@@ -21,11 +23,23 @@ const AuthorsPage = ({
         alert('Loading authors failed ' + error);
       });
     }
+
+    if (courses.length === 0) {
+      actions.loadCourses().catch(error => {
+        alert('Loading courses failed ' + error);
+      });
+    }
   }, [])
 
-  async function handleDeleteAuthor(author) {
-    await actions.deleteAuthor(author)
+  function handleDeleteAuthor(author) {
+    const authorHasCourse = courses.find(course => course.authorId === author.id);
+    if (authorHasCourse) {
+      toast.error('Author can not be deleted. Author has courses.');
+      return;
+    }
+    actions.deleteAuthor(author)
       .then(() => toast.success('Author deleted.'))
+      .then(() => actions.loadAuthors())
       .catch(error => {
         toast.error('Delete failed. ' + error.message, { autoClose: false })
       })
@@ -58,6 +72,7 @@ const AuthorsPage = ({
 }
 
 AuthorsPage.propTypes = {
+  courses: PropTypes.array.isRequired,
   authors: PropTypes.array.isRequired,
   loading: PropTypes.bool.isRequired,
   actions: PropTypes.object.isRequired
@@ -65,6 +80,7 @@ AuthorsPage.propTypes = {
 
 function mapStateToProps(state) {
   return {
+    courses: state.courses,
     authors: state.authors,
     loading: state.apiCallsInProgress > 0
   } 
@@ -74,6 +90,7 @@ function mapDispatchToProps(dispatch) {
   return {
     actions: {
       loadAuthors: bindActionCreators(authorActions.loadAuthors, dispatch),
+      loadCourses: bindActionCreators(coursesActions.loadCourses, dispatch),
       deleteAuthor: bindActionCreators(authorActions.deleteAuthor, dispatch)
     }
   };
